+++ CHANGELOG.md
@@ -12,6 +12,7 @@
     date" warning when initializing a new statistics database.
   * Hardened per-althost content lookups.
   * Precompiled regular expressions for improved performance.
+  * Eliminated data races on console-logging state.
 
 # v1.1.15 (2026-05-07 19:36:24)